Add Lecture 26 split projects
Split projects to eat-go-admin-api, eatgo-common and eatgo-customer-api.
