+++ eatgo-admin-api/src/main/kotlin/kr/co/fastcampus/eatgo/EatgoAdminApiApplication.kt
@@ -1,0 +1,11 @@
+package kr.co.fastcampus.eatgo
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class EatgoAdminApiApplication
+
+fun main(args: Array<String>) {
+    runApplication<EatgoAdminApiApplication>(*args)
+}
