+++ eatgo-admin-api/src/main/kotlin/kr/co/fastcampus/eatgo/interfaces/RestaurantErrorAdvise.kt
@@ -1,0 +1,16 @@
+package kr.co.fastcampus.eatgo.interfaces
+
+import kr.co.fastcampus.eatgo.domain.RestaurantNotFoundException
+import org.springframework.http.HttpStatus
+import org.springframework.web.bind.annotation.ControllerAdvice
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.ResponseBody
+import org.springframework.web.bind.annotation.ResponseStatus
+
+@ControllerAdvice
+class RestaurantErrorAdvise {
+    @ResponseBody
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(RestaurantNotFoundException::class)
+    fun handleNotFound() = "{}"
+}
