+++ eatgo-admin-api/src/test/kotlin/kr/co/fastcampus/eatgo/EatgoAdminApiApplicationTests.kt
@@ -1,0 +1,11 @@
+package kr.co.fastcampus.eatgo
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class EatgoAdminApiApplicationTests {
+    @Test
+    fun contextLoads() {
+    }
+}
