--- eatgo-api/src/main/kotlin/kr/co/fastcampus/eatgo/EatgoApplication.kt
@@ -1,11 +1,0 @@
-package kr.co.fastcampus.eatgo
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class EatGoApplication
-
-fun main(args: Array<String>) {
-    runApplication<EatGoApplication>(*args)
-}
