--- eatgo-api/src/main/kotlin/kr/co/fastcampus/eatgo/domain/ReviewRepository.kt
@@ -1,8 +1,0 @@
-package kr.co.fastcampus.eatgo.domain
-
-import org.springframework.data.repository.CrudRepository
-
-interface ReviewRepository : CrudRepository<Review, Long> {
-    fun save(review: Review): Review
-    fun findAllByRestaurantId(restaurantId: Long): ArrayList<Review>
-}
