--- eatgo-api/src/test/kotlin/kr/co/fastcampus/eatgo/DemoApplicationTests.kt
@@ -1,11 +1,0 @@
-package kr.co.fastcampus.eatgo
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class DemoApplicationTests {
-    @Test
-    fun contextLoads() {
-    }
-}
