+++ eatgo-common/src/main/kotlin/kr/co/fastcampus/eatgo/domain/MenuItem.kt
