+++ eatgo-common/src/main/kotlin/kr/co/fastcampus/eatgo/domain/RestaurantNotFoundException.kt
