+++ eatgo-common/src/main/kotlin/kr/co/fastcampus/eatgo/domain/ReviewRepository.kt
@@ -1,0 +1,11 @@
+package kr.co.fastcampus.eatgo.domain
+
+import org.springframework.data.repository.CrudRepository
+
+interface ReviewRepository : CrudRepository<Review, Long> {
+    override fun findAll(): List<Review>
+
+    fun save(review: Review): Review
+
+    fun findAllByRestaurantId(restaurantId: Long): ArrayList<Review>
+}
