+++ eatgo-common/src/test/kotlin/kr/co/fastcampus/eatgo/domain/RestaurantTests.kt
