+++ eatgo-customer-api/src/main/kotlin/kr.co.fastcampus.eatgo/EatgoCustomerApiApplication.kt
@@ -1,0 +1,11 @@
+package kr.co.fastcampus.eatgo
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class EatgoCustomerApiApplication
+
+fun main(args: Array<String>) {
+    runApplication<EatgoCustomerApiApplication>(*args)
+}
