+++ eatgo-customer-api/src/main/kotlin/kr.co.fastcampus.eatgo/interfaces/RestaurantErrorAdvise.kt
