+++ eatgo-customer-api/src/test/kotlin/kr.co.fastcampus.eatgo/EatgoCustomerApiApplicationTests.kt
@@ -1,0 +1,11 @@
+package kr.co.fastcampus.eatgo
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class EatgoCustomerApiApplicationTests {
+    @Test
+    fun contextLoads() {
+    }
+}
