+++ settings.gradle.kts
@@ -1,2 +1,5 @@
 rootProject.name = "eatgo"
-include("eatgo-api")
+
+include("eatgo-common")
+include("eatgo-admin-api")
+include("eatgo-customer-api")
